mútiples mejoras

+++ src/main/java/es/iesjandula/reaktor/school_manager_server/dtos/AsignaturaConDepartamentoDto.java
@@ -14,10 +14,20 @@ public class AsignaturaConDepartamentoDto
 
     private int plantilla;
 
-    private int horasNecesarias;
+    /**
+     * Horas necesarias por las asignaturas asignadas (como receptor) al departamento (SUM de horas).
+     */
+    private long horasNecesarias;
 
+    /**
+     * Horas disponibles que aporta la plantilla del departamento (plantilla * horas lectivas por profesor).
+     */
     private long horasTotales;
 
+    /**
+     * Desfase = horasTotales - horasNecesarias.
+     * Positivo = sobran horas, 0 = cuadra (cerrado), negativo = faltan horas.
+     */
     private long desfase;
